refactor: file structuring, add jest
