+++ package.json
@@ -6,7 +6,9 @@
     "scripts": {
         "start:dev": "ts-node-dev --respawn src/index.ts",
         "start:prod": "webpack --env mode=production && node build/bundle.js",
-        "build:dev": "webpack --env mode=development"
+        "build:dev": "webpack --env mode=development",
+        "test": "jest --silent",
+        "test:verbose": "jest"
     },
     "keywords": [],
     "author": "",
@@ -16,10 +18,13 @@
         "uuid": "^9.0.1"
     },
     "devDependencies": {
+        "@types/jest": "^29.5.12",
         "@types/node": "^20.11.17",
         "@types/uuid": "^9.0.8",
         "@types/webpack": "^5.28.5",
         "cross-env": "^7.0.3",
+        "jest": "^29.7.0",
+        "ts-jest": "^29.1.2",
         "ts-loader": "^9.5.1",
         "ts-node": "^10.9.2",
         "ts-node-dev": "^2.0.0",
